+++ agp-compatibility/agp-compatibility-7-2/src/main/java/co/elastic/apm/android/agp72/Agp72CompatibilityManager.java
@@ -1,4 +1,4 @@
-package co.elastic.apm.android.agp;
+package co.elastic.apm.android.agp72;
 
 import org.gradle.api.Project;
